Add files via upload
I added this script because its not often but if a new employee joins. Sometimes you need a list of contacts to be able to do your job. with the script you can select from a group and it will create a .vcf file that can either be imported into icloud.com (iOS) or directly onto an android device. Working on fixing this on the next script to allow to have the files sent directly to an email so that the process of grabbing the files is easier. Currently the files are housed still on the netbox server under /opt/netbox/netbox/media vcftoexport.